Add files via upload
Added a 3D viewer so you can now see your networks live in 3D, and with the added benefit is you can see the CONNECTIONS.  Current version has neurons that don't change charge above a threshold be purged, and those that do be " protected" from dying.  This is now part of the cell class.  I may update to also allow pruning of connections that have diminishing weights and are in significant.  This will be easier now that we are using the cell class to keep track of what happens to the cell during training (or a season if you will).