Add files via upload
Made more of the network functions cell autonomous.  Ability to either fix ranges for the network related genes, or have them be autonomous.  Have a  it of lamarcian for gene 6 where it adjusts the range of the values of weights on a distribution based on the number of connections and then updated the gene!